remove Interactive object
