+++ src/main/scala/ldbc/snb/datagen/transformation/transform/RawToBiTransform.scala
@@ -13,7 +13,7 @@ case class RawToBiTransform(mode: BI, simulationStart: Long, simulationEnd: Long
     with Logging {
   log.debug(s"BI Transformation parameters: $mode")
 
-  val bulkLoadThreshold = Interactive.calculateBulkLoadThreshold(mode.bulkloadPortion, simulationStart, simulationEnd)
+  val bulkLoadThreshold = RawToInteractiveTransform.calculateBulkLoadThreshold(mode.bulkloadPortion, simulationStart, simulationEnd)
 
   def batchPeriodFormat(batchPeriod: String) = batchPeriod match {
     case "year"   => "yyyy"
@@ -45,7 +45,7 @@ case class RawToBiTransform(mode: BI, simulationStart: Long, simulationEnd: Long
         .filter(inBatch($"creationDate", batchStart, batchEnd))
         .pipe(batched)
         .select(
-          Seq($"insert_batch_id".as("batch_id")) ++ Interactive.columns(tpe, df.columns).map(qcol): _*
+          Seq($"insert_batch_id".as("batch_id")) ++ RawToInteractiveTransform.columns(tpe, df.columns).map(qcol): _*
         )
         .repartitionByRange($"batch_id")
         .sortWithinPartitions($"creationDate")
@@ -70,7 +70,7 @@ case class RawToBiTransform(mode: BI, simulationStart: Long, simulationEnd: Long
         case (tpe, v) if tpe.isStatic => tpe -> BatchedEntity(v, None, None)
         case (tpe, v) =>
           tpe -> BatchedEntity(
-            Interactive.snapshotPart(tpe, v, bulkLoadThreshold, filterDeletion = false),
+            RawToInteractiveTransform.snapshotPart(tpe, v, bulkLoadThreshold, filterDeletion = false),
             Some(Batched(insertBatchPart(tpe, v, bulkLoadThreshold, simulationEnd), Seq("batch_id"))),
             if (keepImplicitDeletes || v.columns.contains("explicitlyDeleted"))
               Some(Batched(deleteBatchPart(tpe, v, bulkLoadThreshold, simulationEnd), Seq("batch_id")))
